chore(bedrock): clean up derive package, move test util out or polish for public use, split payload_attributes and improve file names
