+++ op-node/rollup/derive/deposits.go
@@ -1,0 +1,46 @@
+package derive
+
+import (
+	"fmt"
+
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/common/hexutil"
+	"github.com/ethereum/go-ethereum/core/types"
+)
+
+// UserDeposits transforms the L2 block-height and L1 receipts into the transaction inputs for a full L2 block
+func UserDeposits(receipts []*types.Receipt, depositContractAddr common.Address) ([]*types.DepositTx, []error) {
+	var out []*types.DepositTx
+	var errs []error
+
+	for i, rec := range receipts {
+		if rec.Status != types.ReceiptStatusSuccessful {
+			continue
+		}
+		for j, log := range rec.Logs {
+			if log.Address == depositContractAddr && len(log.Topics) > 0 && log.Topics[0] == DepositEventABIHash {
+				dep, err := UnmarshalDepositLogEvent(log)
+				if err != nil {
+					errs = append(errs, fmt.Errorf("malformatted L1 deposit log in receipt %d, log %d: %w", i, j, err))
+				} else {
+					out = append(out, dep)
+				}
+			}
+		}
+	}
+	return out, errs
+}
+
+func DeriveDeposits(receipts []*types.Receipt, depositContractAddr common.Address) ([]hexutil.Bytes, []error) {
+	userDeposits, errs := UserDeposits(receipts, depositContractAddr)
+	encodedTxs := make([]hexutil.Bytes, 0, len(userDeposits))
+	for i, tx := range userDeposits {
+		opaqueTx, err := types.NewTx(tx).MarshalBinary()
+		if err != nil {
+			errs = append(errs, fmt.Errorf("failed to encode user tx %d", i))
+		} else {
+			encodedTxs = append(encodedTxs, opaqueTx)
+		}
+	}
+	return encodedTxs, errs
+}
