+++ op-node/testutils/deposits.go
@@ -1,0 +1,65 @@
+package testutils
+
+import (
+	"math/big"
+	"math/rand"
+
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/core/types"
+)
+
+func GenerateAddress(rng *rand.Rand) (out common.Address) {
+	rng.Read(out[:])
+	return
+}
+
+func RandETH(rng *rand.Rand, max int64) *big.Int {
+	x := big.NewInt(rng.Int63n(max))
+	x = new(big.Int).Mul(x, big.NewInt(1e18))
+	return x
+}
+
+// Returns a DepositEvent customized on the basis of the id parameter.
+func GenerateDeposit(sourceHash common.Hash, rng *rand.Rand) *types.DepositTx {
+	dataLen := rng.Int63n(10_000)
+	data := make([]byte, dataLen)
+	rng.Read(data)
+
+	var to *common.Address
+	if rng.Intn(2) == 0 {
+		x := GenerateAddress(rng)
+		to = &x
+	}
+	var mint *big.Int
+	if rng.Intn(2) == 0 {
+		mint = RandETH(rng, 200)
+	}
+
+	dep := &types.DepositTx{
+		SourceHash: sourceHash,
+		From:       GenerateAddress(rng),
+		To:         to,
+		Value:      RandETH(rng, 200),
+		Gas:        uint64(rng.Int63n(10 * 1e6)), // 10 M gas max
+		Data:       data,
+		Mint:       mint,
+	}
+	return dep
+}
+
+// Generates an EVM log entry with the given topics and data.
+func GenerateLog(addr common.Address, topics []common.Hash, data []byte) *types.Log {
+	return &types.Log{
+		Address: addr,
+		Topics:  topics,
+		Data:    data,
+		Removed: false,
+
+		// ignored (zeroed):
+		BlockNumber: 0,
+		TxHash:      common.Hash{},
+		TxIndex:     0,
+		BlockHash:   common.Hash{},
+		Index:       0,
+	}
+}
